# gomino-src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  community.BanRequest:
    properties:
      note:
        properties:
          content:
            type: string
        type: object
    type: object
  community.CreateCommunityRequest:
    properties:
      icon:
        type: string
      keywords:
        items:
          type: string
        type: array
      name:
        type: string
      primaryLanguage:
        type: string
      privacyMode:
        description: '0: Open, 1: Approval Required, 2: Private'
        type: integer
      tagline:
        type: string
      themeColor:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      api:message:
        type: string
      api:statuscode:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gomino.com
    name: API Support
  description: This is the API server for GoMino.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GoMino API
  version: "1.0"
paths:
  /g/s-x{comId}/community:
    get:
      consumes:
      - application/json
      description: Get details of a specific community
      parameters:
      - description: Community ID or NDC ID
        in: path
        name: comId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get community details
      tags:
      - community
...
  /g/s/community:
    post:
      consumes:
      - application/json
      description: Create a new community with the given details
      parameters:
      - description: Community details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/community.CreateCommunityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new community
      tags:
      - community
  /g/s/community/joined:
    get:
      consumes:
      - application/json
      description: Get a list of communities the user has joined
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get joined communities
      tags:
      - community
...
  /x{comId}/s/admin/operation:
    get:
      consumes:
      - application/json
      description: Get moderation logs for the community
      parameters:
      - description: Community ID (NDC ID)
        in: path
        name: comId
        required: true
        type: string
      - description: Page size (max 100)
        in: query
        name: size
        type: integer
      - description: Filter by Object ID
        in: query
        name: objectId
        type: string
      - description: Filter by Object Type
        in: query
        name: objectType
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get moderation history
      tags:
      - community-admin
...
  /x{comId}/s/community/join:
    post:
      consumes:
      - application/json
      description: Join a specific community
      parameters:
      - description: Community ID (NDC ID)
        in: path
        name: comId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Join a community
      tags:
      - community
  /x{comId}/s/community/leave:
    post:
      consumes:
      - application/json
      description: Leave a specific community
      parameters:
      - description: Community ID (NDC ID)
        in: path
        name: comId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Leave a community
      tags:
      - community
...
  /x{comId}/s/user-profile/{userId}/ban:
    post:
      consumes:
      - application/json
      description: Ban a user from the community
      parameters:
      - description: Community ID (NDC ID)
        in: path
        name: comId
        required: true
        type: string
      - description: Target User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Ban reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/community.BanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Ban user
      tags:
      - community-admin
...
  /x{comId}/s/user-profile/{userId}/curator:
    post:
      consumes:
      - application/json
      description: Promote a user to curator role
      parameters:
      - description: Community ID (NDC ID)
        in: path
        name: comId
        required: true
        type: string
      - description: Target User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Promote user to curator
      tags:
      - community-admin
...
  /x{comId}/s/user-profile/{userId}/leader:
    post:
      consumes:
      - application/json
      description: Promote a user to leader role
      parameters:
      - description: Community ID (NDC ID)
        in: path
        name: comId
        required: true
        type: string
      - description: Target User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Promote user to leader
      tags:
      - community-admin
...
  /x{comId}/s/user-profile/{userId}/transfer-agent:
    post:
      consumes:
      - application/json
      description: Transfer agent status to another user
      parameters:
      - description: Community ID (NDC ID)
        in: path
        name: comId
        required: true
        type: string
      - description: Target User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Transfer agent status
      tags:
      - community-admin
  /x{comId}/s/user-profile/{userId}/unban:
    post:
      consumes:
      - application/json
      description: Unban a user from the community
      parameters:
      - description: Community ID (NDC ID)
        in: path
        name: comId
        required: true
        type: string
      - description: Target User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Unban reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/community.BanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Unban user
      tags:
      - community-admin
swagger: "2.0"